RECOVERY: Prepare recovery database to hold NLM/NSM information also
This commit doesn't actually implement any new recovery database records, it
just sets the stage for their addition as necessary.

Mostly this will just recognize them as NOT NFSv4.x clientid entries and issue
a warning message should they appear in the database (but otherwise they will
be ignored and left alone).

+++ src/SAL/recovery/recovery_fs.c
@@ -54,7 +54,8 @@ static char *reclaim_complete_marker = "reclaim_complete";
  * @return length of path or -ve on error
  */
 static int fs_make_ip_recov_dir_name(sockaddr_t *saddr, bool is_old,
-				     int path_size, char *path)
+				     int path_size, char *path,
+				     enum log_components component)
 {
 	int rc;
 	char ipstring[SOCK_NAME_MAX];
@@ -62,7 +63,7 @@ static int fs_make_ip_recov_dir_name(sockaddr_t *saddr, bool is_old,
 
 	display_sockip(&dspbuf, saddr);
 
-	LogDebug(COMPONENT_CLIENTID, "ip_based=%d addr_int=%s is_old=%d",
+	LogDebug(component, "ip_based=%d addr_int=%s is_old=%d",
 		 nfs_param.nfsv4_param.recovery_backend_ipbased, ipstring,
 		 is_old);
 	if (is_old) {
@@ -76,15 +77,15 @@ static int fs_make_ip_recov_dir_name(sockaddr_t *saddr, bool is_old,
 	}
 
 	if (unlikely(rc >= path_size)) {
-		LogCrit(COMPONENT_CLIENTID, "Path %s too long", path);
+		LogCrit(component, "Path %s too long", path);
 		return -1;
 	} else if (unlikely(rc < 0)) {
-		LogCrit(COMPONENT_CLIENTID,
+		LogCrit(component,
 			"Unexpected return from snprintf %d error %s (%d)", rc,
 			strerror(errno), errno);
 		return rc;
 	}
-	LogDebug(COMPONENT_CLIENTID, "dir = %s", path);
+	LogDebug(component, "dir = %s", path);
 	return rc;
 }
 
@@ -113,7 +114,7 @@ int fs_create_recov_dir(void)
 	if (err == -1 && errno != EEXIST) {
 		err = errno;
 		LogCrit(COMPONENT_CLIENTID,
-			"Failed to create v4 recovery dir (%s), errno: %s (%d)",
+			"Failed to create Recovery Directory (%s), errno: %s (%d)",
 			nfs_param.nfsv4_param.recov_root, strerror(errno),
 			errno);
 		return -err;
@@ -129,7 +130,7 @@ int fs_create_recov_dir(void)
 
 	if (v4_recov_dir_len >= sizeof(v4_recov_dir)) {
 		LogCrit(COMPONENT_CLIENTID,
-			"v4 recovery dir path (%s/%s) is too long",
+			"Recovery Directory path (%s/%s) is too long",
 			nfs_param.nfsv4_param.recov_root,
 			nfs_param.nfsv4_param.recov_dir);
 		return -ENAMETOOLONG;
@@ -141,13 +142,14 @@ int fs_create_recov_dir(void)
 	       dir_len + 1);
 	dir_len = 1 + root_len + dir_len;
 
-	LogDebug(COMPONENT_CLIENTID, "v4_recov_dir=%s", v4_recov_dir);
+	LogInfo(COMPONENT_CLIENTID, "Recovery Directory = %s", v4_recov_dir);
 
 	err = mkdir(v4_recov_dir, 0755);
+
 	if (err == -1 && errno != EEXIST) {
 		err = errno;
 		LogCrit(COMPONENT_CLIENTID,
-			"Failed to create v4 recovery dir(%s), errno: %s (%d)",
+			"Failed to create Recovery Directory (%s), errno: %s (%d)",
 			v4_recov_dir, strerror(errno), errno);
 		return -err;
 	}
@@ -162,7 +164,7 @@ int fs_create_recov_dir(void)
 
 	if (v4_old_dir_len >= sizeof(v4_old_dir)) {
 		LogCrit(COMPONENT_CLIENTID,
-			"v4 recovery dir path (%s/%s) is too long",
+			"Old Recovery Directory path (%s/%s) is too long",
 			nfs_param.nfsv4_param.recov_root,
 			nfs_param.nfsv4_param.recov_old_dir);
 		return -ENAMETOOLONG;
@@ -174,14 +176,14 @@ int fs_create_recov_dir(void)
 	       old_len + 1);
 	old_len = 1 + root_len + old_len;
 
-	LogDebug(COMPONENT_CLIENTID, "v4_old_dir=%s", v4_old_dir);
+	LogInfo(COMPONENT_CLIENTID, "Old Recovery Directory = %s", v4_old_dir);
 
 	err = mkdir(v4_old_dir, 0755);
 
 	if (err == -1 && errno != EEXIST) {
 		err = errno;
 		LogCrit(COMPONENT_CLIENTID,
-			"Failed to create v4 recovery dir(%s), errno: %s (%d)",
+			"Failed to createRecovery Directory (%s), errno: %s (%d)",
 			v4_old_dir, strerror(errno), errno);
 		return -err;
 	}
@@ -194,15 +196,17 @@ int fs_create_recov_dir(void)
 		memcpy(v4_recov_dir + dir_len, node, node_size + 1);
 		memcpy(v4_old_dir + old_len, node, node_size + 1);
 
-		LogDebug(COMPONENT_CLIENTID, "v4_recov_dir=%s", v4_recov_dir);
-		LogDebug(COMPONENT_CLIENTID, "v4_old_dir=%s", v4_old_dir);
+		LogInfo(COMPONENT_CLIENTID, "Recovery Directory = %s",
+			v4_recov_dir);
+		LogInfo(COMPONENT_CLIENTID, "Old Recovery Directory = %s",
+			v4_old_dir);
 
 		err = mkdir(v4_recov_dir, 0755);
 
 		if (err == -1 && errno != EEXIST) {
 			err = errno;
 			LogCrit(COMPONENT_CLIENTID,
-				"Failed to create v4 recovery dir(%s), errno: %s (%d)",
+				"Failed to create Recovery Directory dir(%s), errno: %s (%d)",
 				v4_recov_dir, strerror(errno), errno);
 			return -err;
 		}
@@ -212,16 +216,11 @@ int fs_create_recov_dir(void)
 		if (err == -1 && errno != EEXIST) {
 			err = errno;
 			LogCrit(COMPONENT_CLIENTID,
-				"Failed to create v4 recovery dir(%s), errno: %s (%d)",
+				"Failed to create Old Recovery Directory (%s), errno: %s (%d)",
 				v4_old_dir, strerror(errno), errno);
 			return -err;
 		}
 	}
-
-	LogInfo(COMPONENT_CLIENTID, "NFSv4 Recovery Directory %s",
-		v4_recov_dir);
-	LogInfo(COMPONENT_CLIENTID, "NFSv4 Recovery Directory (old) %s",
-		v4_old_dir);
 
 	LogEvent(COMPONENT_CLIENTID,
 		 "fs recovery backend initialization complete");
@@ -239,20 +238,22 @@ int fs_create_recov_dir(void)
  *
  * @return Length of path or -errno if failure
  */
-int fs_make_clid_path(const char *name, sockaddr_t *serv, char *path, bool make)
+int fs_make_path(const char *name, sockaddr_t *serv, char *path, int pathsize,
+		 bool make, enum log_components component)
 {
 	int err = 0;
 	int length, srcpos, dstpos;
 
-	if (nfs_param.nfsv4_param.recovery_backend_ipbased) {
-		dstpos = fs_make_ip_recov_dir_name(serv, false, PATH_MAX, path);
+	if (nfs_param.nfsv4_param.recovery_backend_ipbased && serv != NULL) {
+		dstpos = fs_make_ip_recov_dir_name(serv, false, PATH_MAX, path,
+						   component);
 
 		if (make) {
 			err = mkdir(path, 0700);
 			if (err == -1 && errno != EEXIST) {
 				err = errno;
 				LogEvent(
-					COMPONENT_CLIENTID,
+					component,
 					"Failed to create recov_dir (%s), errno: %s (%d)",
 					path, strerror(err), err);
 				return -err;
@@ -263,13 +264,19 @@ int fs_make_clid_path(const char *name, sockaddr_t *serv, char *path, bool make)
 		dstpos = v4_recov_dir_len;
 	}
 
-	if (isMidDebug(COMPONENT_CLIENTID)) {
+	path[dstpos] = '\0';
+
+	if (isMidDebug(component)) {
 		char ipstring[SOCK_NAME_MAX];
 		struct display_buffer dspbuf = { sizeof(ipstring), ipstring,
 						 ipstring };
-		display_sockip(&dspbuf, serv);
-		LogDebug(COMPONENT_CLIENTID, "server_addr=%s recov_dir=%s",
-			 ipstring, path);
+		if (serv != NULL)
+			display_sockip(&dspbuf, serv);
+		else
+			display_cat(&dspbuf, "<NONE>");
+
+		LogDebug(component, "server_addr=%s recov_dir=%s", ipstring,
+			 path);
 	}
 
 	/* break clientid down if it is greater than max dir name
@@ -287,11 +294,15 @@ int fs_make_clid_path(const char *name, sockaddr_t *serv, char *path, bool make)
 		path[dstpos++] = '/';
 
 		/* Make sure there's still room in path */
-		if ((dstpos + len) >= sizeof(path))
+		if ((dstpos + len) >= pathsize) {
+			LogDebug(component, "no room in path for name %s",
+				 name);
 			return -ENOMEM;
+		}
 
 		if (len <= NAME_MAX) {
 			memcpy(path + dstpos, name + srcpos, len + 1);
+			dstpos += len;
 			if (make) {
 				err = mkdir(path, 0700);
 				/* Handle error outside loop */
@@ -318,14 +329,14 @@ int fs_make_clid_path(const char *name, sockaddr_t *serv, char *path, bool make)
 	if (err == -1 && errno != EEXIST) {
 		err = errno;
 		LogEvent(
-			COMPONENT_CLIENTID,
-			"Failed to create client in recovery dir (%s), errno: %s (%d)",
+			component,
+			"Failed to create entry in recovery dir (%s), errno: %s (%d)",
 			path, strerror(err), err);
 		return -err;
 	}
 
-	LogDebug(COMPONENT_CLIENTID, "%srecovery entry dir [%s]",
-		 make ? "Created " : "", path);
+	LogDebug(component, "%srecovery entry dir [%s] dstpos=%d",
+		 make ? "Created " : "", path, dstpos);
 
 	return dstpos;
 }
@@ -337,9 +348,9 @@ void fs_add_clid(nfs_client_id_t *clientid)
 
 	clientid->cid_recov_tag = nfs4_create_clid_name(clientid, NULL);
 
-	err = fs_make_clid_path(clientid->cid_recov_tag,
-				&clientid->cid_client_record->cr_server_addr,
-				path, true);
+	err = fs_make_path(clientid->cid_recov_tag,
+			   &clientid->cid_client_record->cr_server_addr, path,
+			   sizeof(path), true, COMPONENT_CLIENTID);
 
 	assert(err > 0);
 }
@@ -368,9 +379,9 @@ void fs_reclaim_complete(nfs_client_id_t *clientid)
 	marker_len = strlen(reclaim_complete_marker);
 
 	/* Fill in path from the cid_recov_tag, don't create directories */
-	dstpos = fs_make_clid_path(clientid->cid_recov_tag,
-				   &clientid->cid_client_record->cr_server_addr,
-				   path, false);
+	dstpos = fs_make_path(clientid->cid_recov_tag,
+			      &clientid->cid_client_record->cr_server_addr,
+			      path, sizeof(path), false, COMPONENT_CLIENTID);
 
 	assert(dstpos > 0);
 
@@ -407,7 +418,7 @@ void fs_reclaim_complete(nfs_client_id_t *clientid)
  *
  * @param[in] path The path of the client-id on the stable storage.
  */
-static void fs_rm_client_records(char *path)
+static void fs_rm_client_records(char *path, enum log_components component)
 {
 	DIR *dp;
 	struct dirent *dentp;
@@ -418,8 +429,8 @@ static void fs_rm_client_records(char *path)
 
 	if (dp == NULL) {
 		int err = errno;
-		LogEvent(COMPONENT_CLIENTID, "opendir %s failed errno: %s (%d)",
-			 path, strerror(err), err);
+		LogEvent(component, "opendir %s failed errno: %s (%d)", path,
+			 strerror(err), err);
 		return;
 	}
 
@@ -436,16 +447,16 @@ static void fs_rm_client_records(char *path)
 			      dentp->d_name);
 
 		if (unlikely(rc >= sizeof(del_path))) {
-			LogCrit(COMPONENT_CLIENTID, "Path %s/%s too long", path,
+			LogCrit(component, "Path %s/%s too long", path,
 				dentp->d_name);
 		} else if (unlikely(rc < 0)) {
 			err = errno;
-			LogCrit(COMPONENT_CLIENTID,
+			LogCrit(component,
 				"Unexpected return from snprintf %d error %s (%d)",
 				rc, strerror(err), err);
 		} else if (unlink(del_path) < 0) {
 			err = errno;
-			LogEvent(COMPONENT_CLIENTID,
+			LogEvent(component,
 				 "unlink of %s failed errno: %s (%d)", del_path,
 				 strerror(err), err);
 		}
@@ -454,24 +465,24 @@ static void fs_rm_client_records(char *path)
 	(void)closedir(dp);
 }
 
-static void fs_rm_clid_impl(int position, char *recov_dir, int len,
-			    char *parent_path, int parent_len)
+static int fs_rm_entry_impl(int position, char *recov_dir, int len,
+			    char *parent_path, int parent_len,
+			    enum log_components component)
 {
 	int err;
 	char *path;
 	int segment_len;
 	int total_len;
 
-	LogDebug(COMPONENT_CLIENTID,
-		 "position=%d len=%d  parent_path=%s recov_dir=%s", position,
-		 len, parent_path, recov_dir);
+	LogDebug(component, "position=%d len=%d  parent_path=%s recov_dir=%s",
+		 position, len, parent_path, recov_dir);
 
 	if (position == len) {
 		/* We are at the tail directory of the clid,
 		* remove revoked handles and reclaim complete marker, if any.
 		*/
-		fs_rm_client_records(parent_path);
-		return;
+		fs_rm_client_records(parent_path, component);
+		return true;
 	}
 
 	if ((len - position) > NAME_MAX)
@@ -495,26 +506,32 @@ static void fs_rm_clid_impl(int position, char *recov_dir, int len,
 	 * be the same as len, and this recursion will not process an additional
 	 * path component, rather it will clean the final directory.
 	 */
-	fs_rm_clid_impl(position + segment_len, recov_dir, len, path,
-			total_len - 1);
+	err = fs_rm_entry_impl(position + segment_len, recov_dir, len, path,
+			       total_len - 1, component);
+	if (err == -1) {
+		gsh_free(path);
+		return err;
+	}
 
 	/* Remove this sub-directory on tail recursion */
 
-	LogDebug(COMPONENT_CLIENTID, "Will remove %s", path);
+	LogDebug(component, "Will remove %s", path);
 
 	err = rmdir(path);
 
 	if (err == -1) {
-		err = errno;
+		int error = errno;
 		LogEvent(
-			COMPONENT_CLIENTID,
+			component,
 			"Failed to remove client recovery dir (%s), errno: %s (%d)",
-			path, strerror(err), err);
+			path, strerror(error), error);
 	} else {
-		LogDebug(COMPONENT_CLIENTID, "Removed client dir (%s)", path);
+		LogDebug(component, "Removed client dir (%s)", path);
 	}
 
 	gsh_free(path);
+
+	return err;
 }
 
 void fs_rm_clid(nfs_client_id_t *clientid)
@@ -541,7 +558,7 @@ void fs_rm_clid(nfs_client_id_t *clientid)
 
 		ip_recov_dir_len = fs_make_ip_recov_dir_name(
 			&clientid->cid_client_record->cr_server_addr, false,
-			sizeof(ip_recov_dir), ip_recov_dir);
+			sizeof(ip_recov_dir), ip_recov_dir, COMPONENT_CLIENTID);
 
 		/*
 		   In the case of v4.1, confirmed client ids are removed as
@@ -570,9 +587,10 @@ void fs_rm_clid(nfs_client_id_t *clientid)
 					 client_str);
 			}
 
-			fs_rm_clid_impl(0, clientid->cid_recov_tag,
-					strlen(clientid->cid_recov_tag),
-					ip_recov_dir, ip_recov_dir_len);
+			fs_rm_entry_impl(0, clientid->cid_recov_tag,
+					 strlen(clientid->cid_recov_tag),
+					 ip_recov_dir, ip_recov_dir_len,
+					 COMPONENT_CLIENTID);
 			gsh_free(clientid->cid_recov_tag);
 			clientid->cid_recov_tag = NULL;
 		} else if (clientid->cid_minorversion > 0 &&
@@ -588,9 +606,9 @@ void fs_rm_clid(nfs_client_id_t *clientid)
 				 ip_recov_dir);
 		}
 	} else {
-		fs_rm_clid_impl(0, clientid->cid_recov_tag,
-				strlen(clientid->cid_recov_tag), v4_recov_dir,
-				v4_recov_dir_len);
+		fs_rm_entry_impl(0, clientid->cid_recov_tag,
+				 strlen(clientid->cid_recov_tag), v4_recov_dir,
+				 v4_recov_dir_len, COMPONENT_CLIENTID);
 		gsh_free(clientid->cid_recov_tag);
 		clientid->cid_recov_tag = NULL;
 	}
@@ -735,6 +753,21 @@ bool process_recovery_entry(char *build_clid, char *tgtdir, int takeover,
 	int len, cid_len;
 	bool reclaim_complete;
 	clid_entry_t *new_ent;
+	enum recovery_type recovery_type = entry_recovery_type(build_clid);
+	enum log_components component;
+
+	if (recovery_type == NFS4_CLID_ENTRY)
+		component = COMPONENT_CLIENTID;
+	else
+		component = COMPONENT_NLM;
+
+	LogFullDebug(component, "Processing entry %s type %d", build_clid,
+		     recovery_type);
+
+	if (recovery_type > NFS4_CLID_ENTRY) {
+		/* Process an NLM recovery entry */
+		return parse_nlm_entry(build_clid, recovery_type);
+	}
 
 	/* Process an NFSv4 clid entry
 	 *
@@ -772,7 +805,7 @@ bool process_recovery_entry(char *build_clid, char *tgtdir, int takeover,
 		reclaim_complete = fs_check_reclaim_complete(sub_path);
 		new_ent = nfs4_add_clid_entry(build_clid, reclaim_complete);
 		fs_cp_pop_revoked_delegs(new_ent, sub_path, tgtdir, !takeover);
-		LogDebug(COMPONENT_CLIENTID,
+		LogDebug(component,
 			 "added %s to clid list, reclaim_complete %d",
 			 new_ent->cl_name, reclaim_complete);
 	}
@@ -903,7 +936,8 @@ static int fs_read_recov_clids_impl(const char *parent_path, char *clid_str,
 				LogDebug(COMPONENT_CLIENTID, "Would remove %s",
 					 sub_path);
 			} else {
-				fs_rm_client_records(sub_path);
+				fs_rm_client_records(sub_path,
+						     COMPONENT_CLIENTID);
 
 				rc = rmdir(sub_path);
 
@@ -913,6 +947,11 @@ static int fs_read_recov_clids_impl(const char *parent_path, char *clid_str,
 						"Failed to rmdir (%s), errno: %s (%d)",
 						sub_path, strerror(errno),
 						errno);
+				} else {
+					LogFullDebugAlt(COMPONENT_CLIENTID,
+							COMPONENT_NLM,
+							"Succeeded rmdir (%s)",
+							sub_path);
 				}
 			}
 		}
@@ -971,14 +1010,14 @@ void fs_read_recov_clids_takeover(nfs_grace_start_t *gsp)
 	char recov_dir[PATH_MAX];
 	char old_dir[PATH_MAX];
 
-	LogDebug(COMPONENT_CLIENTID, "ip_based=%d, gsp=%p",
-		 nfs_param.nfsv4_param.recovery_backend_ipbased, gsp);
+	LogDebugAlt(COMPONENT_CLIENTID, COMPONENT_NLM, "ip_based=%d, gsp=%p",
+		    nfs_param.nfsv4_param.recovery_backend_ipbased, gsp);
 	if (nfs_param.nfsv4_param.recovery_backend_ipbased) {
 		if (gsp) {
 			/* should always be set */
 			if (gsp->ipaddr) {
-				LogDebug(COMPONENT_CLIENTID, "gsp_ipaddr=%s",
-					 gsp->ipaddr);
+				LogDebugAlt(COMPONENT_CLIENTID, COMPONENT_NLM,
+					    "gsp_ipaddr=%s", gsp->ipaddr);
 				rc = ip_str_to_sockaddr(gsp->ipaddr, &saddr);
 				if (rc != 0) {
 					LogWarn(COMPONENT_CLIENTID,
@@ -989,23 +1028,25 @@ void fs_read_recov_clids_takeover(nfs_grace_start_t *gsp)
 
 				fs_make_ip_recov_dir_name(&saddr, false,
 							  sizeof(recov_dir),
-							  recov_dir);
+							  recov_dir,
+							  COMPONENT_CLIENTID);
 				create_dir(recov_dir);
 				fs_make_ip_recov_dir_name(&saddr, true,
 							  sizeof(old_dir),
-							  old_dir);
+							  old_dir,
+							  COMPONENT_CLIENTID);
 				create_dir(old_dir);
 			}
 		} else {
-			LogDebug(
-				COMPONENT_CLIENTID,
+			LogDebugAlt(
+				COMPONENT_CLIENTID, COMPONENT_NLM,
 				"IP BASED v4_recov_dir_len=%d strlen(v4_recov_dir)=%ld",
 				v4_recov_dir_len, strlen(v4_recov_dir));
 			memcpy(recov_dir, v4_recov_dir, v4_recov_dir_len + 1);
 			memcpy(old_dir, v4_old_dir, v4_old_dir_len + 1);
 		}
-		LogDebug(COMPONENT_CLIENTID, "IP BASED END %s %s", recov_dir,
-			 old_dir);
+		LogDebugAlt(COMPONENT_CLIENTID, COMPONENT_NLM,
+			    "IP BASED END %s %s", recov_dir, old_dir);
 	} else {
 		memcpy(recov_dir, v4_recov_dir, v4_recov_dir_len + 1);
 		memcpy(old_dir, v4_old_dir, v4_old_dir_len + 1);
@@ -1123,7 +1164,8 @@ void fs_clean_old_recov_dir_impl(char *parent_path)
 		if (dentp->d_type == DT_REG) {
 			/* Remove regular files: revoked file handles */
 			/* and reclaim_complete marker now */
-			LogDebug(COMPONENT_CLIENTID, "Will remove %s", path);
+			LogDebugAlt(COMPONENT_CLIENTID, COMPONENT_NLM,
+				    "Will remove %s", path);
 
 			if (unlink(path) < 0) {
 				LogEvent(COMPONENT_CLIENTID,
@@ -1141,6 +1183,10 @@ void fs_clean_old_recov_dir_impl(char *parent_path)
 				LogEvent(COMPONENT_CLIENTID,
 					 "Failed to remove %s, errno: %s (%d)",
 					 path, strerror(errno), errno);
+			} else {
+				LogFullDebugAlt(COMPONENT_CLIENTID,
+						COMPONENT_NLM,
+						"Succeeded rmdir (%s)", path);
 			}
 		} else {
 			LogEvent(COMPONENT_CLIENTID,
